+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Unreleased
 
+* Escape newlines etc.
+* Reorder columns, swapping `new_value` & `old_value`, & `new_version` & `old_version`
+* New `id` column format, merging object type & id
+
 # v0.1.0 - 6 Jan 2020
 
 * Initial version